Enhancements
Long story short I changed a bunch of things to make the model perform much more reliably in all aspects.

Memory has greater retention and  interacts with both the fast and coder model.

You can make code changes on the fly if not happy with prior output from the model.

You have a personal history with the model as far as a "human relationship" goes.